Update to reflect decisions for BD and OP boxes
Removed the note about separating out and storing born digital materials in BD boxes. We may want to add an instruction about when it is appropriate to store BD in a separate box (but still using a regular box number). Updated instructions regarding oversized papers, OBVs, and BVs, but it was challenging and could probably use further revision. Not sure what other sections may refer to how we identify these containers and may need edits.